rectification 19

+++ app/Http/Controllers/ProduitController.php
@@ -3,11 +3,14 @@
 namespace App\Http\Controllers;
 
 use App\Models\Produit;
+use App\Models\Commande;
 use App\Models\Categorie;
 use App\Models\Fournisseur;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Session;
 use App\Http\Requests\StoreProduitRequest;
 use App\Http\Requests\UpdateProduitRequest;
+use App\Models\Commande_detail;
 
 class ProduitController extends Controller
 {
@@ -83,9 +86,14 @@ class ProduitController extends Controller
      */
     public function update(UpdateProduitRequest $request, $id)
     {
+        $user = Auth::user();
         $produit = Produit::find($id);
         $produit->update($request->all());
-        return redirect()->route('produit.index');
+        if($user->hasRole('Magasinier')){
+            return redirect('/magasinier');
+        }else{
+            return redirect()->route('produit.index');
+        }
     }
 
     /**
@@ -94,7 +102,13 @@ class ProduitController extends Controller
     public function destroy($id)
     {
         $produit = Produit::find($id);
+        $commande = Commande_detail::where('produit_id',$id)->first();
+        if($commande){
+            return back()->withErrors(['message' => 'Impossible de supprimer le produit car il est associé à une commande.']);
+        }
+
         $produit->delete();
+        Session::flash('success', 'Produit supprimé avec succès.');
         return back();
     }
 }
